plot: allow setting the image size and the format through REST
Currently, only format recognized is SVG, but it's extensible to other formats easily like PNG, JPEG, etc.
The image width option + height are set so that if only the width is set, then the height is determined from the golden ratio, and if the width isn't set, it defaults to 4 inches.
The size settings are parsed with units, so it works with in, cm, mm, etc.