edgex: dynamically determine the address to use in edgex registration
We need to dynamically figure this out, otherwise the address specified on the command line, such as "", "0.0.0.0", "127.0.0.7", "localhost", etc. won't be usable by edgex unless edgex is running on the same machine. This way, we actually make a network connection to edgex and then see what IP address we used in that connection and provide that to edgex. This always ensures that if the two machines are reachable on the network you only have to specify the address of where to find edgex and not where to find this server.