http: use existing influx client in all go routines
We now create one influx http client, and use that same one in all go routines, as it is guaranteed to be threadsafe. This ensures that the handler doesn't need to care how to setup the client, and details such as where to find influx and such. We also provide the batch points config to all handlers as well, so they don't have to allocate that or care how it's setup.
Also, now we log + time all HTTP operations.