cmd/influxproxy/main.go: use upstream configurator pkg, add "set" impl
Rather than use the config package in this repo, we can use the
generalized one at github.com/configurator/tomlconfigurator.
This requires moving the struct definition into this file, and also
changing the implementation of the various config commands to use the
tomlconfigurator package.

Using the upstream package also allows us to implement the set config
command to update a config item in the file from the command line.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>