many: replace glide with go modules to build snap
The delhi branch of edgex-go now has an empty module file under
/cmd/config-seed/res/properties to ignore that folder as there are files
there with ";" in them, which go mod explicitly forbids.

This also means we can get rid of glide.{yaml,lock} and stop building
with glide in the snap and coincidentally now support all architectures
as before with glide we were limited by what architectures glide was
available for.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>