go.mod, snapcraft: build w/ 1.15, update + pin edgex deps
For whatever reason, there is a problem using the most up to date versions of
the edgex go modules, using go-mod-core-contracts v0.1.149 and go-mod-bootstrap
v0.0.72 fails to build together. So instead pin go-mod-core-contracts to an
older version so that it builds properly.

Also update the snapcraft.yaml to use go as a snap with build-snaps.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>